Update on "[quant][fix] Compare resnet with quantizer api with the prepare_fx and decomposed convert flow"
Summary:
Using a decomposed convert to make sure we get exact match, this means the nodes in resnet are
annotated correctly

Test Plan:
python test/test_quantization.py TestQuantizePT2EModels.test_resnet18_with_quantizer_api

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
